+++ src/accounts/index.ts
@@ -1,2 +1,2 @@
-export { getMinimumReserve } from './keypair';
+export { generateKeypair, getMinimumReserve } from './keypair';
 export { AccountManager } from './manager';